+++ src/D3D11Renderer/Rendering/Scenes/Game/GameSceneConsumer.cs
@@ -52,6 +52,7 @@ namespace Sovereign.D3D11Renderer.Rendering.Scenes.Game
         {
             /* Handle general processing. */
             scene.PopulateBuffers(gameResourceManager.VertexBuffer.Buffer,
+                gameResourceManager.IndexBuffer.Buffer,
                 gameResourceManager.DrawBuffer, 
                 out var drawCount);
             gameResourceManager.DrawCount = drawCount;